refactored Validations for views

+++ app/routes/admin.py
@@ -3,7 +3,8 @@ from flask_restful import Resource, Api, reqparse
 from flask_jwt_extended import jwt_required
 from app import db
 from app.models import Category, Question
-from app.schema import CategorySchema, CategoryUpdateSchema, QuestionCreateSchema, QuestionListSchema, QuestionUpdateSchema
+from app.schema import CategoryListSchema, CategoryCreateUpdateSchema, QuestionCreateSchema, QuestionListSchema, \
+    QuestionUpdateSchema
 from marshmallow import ValidationError
 
 
@@ -11,35 +12,79 @@ class CategoryResource(Resource):
 
     @jwt_required()
     def get(self):
+        """
+        Get list of all categories
+        """
         cateogries = Category.query.all()
-        categories_list = CategorySchema(many=True).dump(cateogries)
+        categories_list = CategoryListSchema(many=True).dump(cateogries)
         return {"categories": categories_list}
 
     @jwt_required()
     def post(self):
+        """
+        Create a new category.
+        ---
+        parameters:
+          - name: name
+            in: request data
+            required: true
+            type: string
+        """
         try:
-            category_schema = CategoryUpdateSchema().load(request.json)
+            category_schema = CategoryCreateUpdateSchema().load(request.json)
             category_obj = Category(**category_schema)
             db.session.add(category_obj)
             db.session.commit()
         except ValidationError as err:
             return err.messages, 400
-        return CategorySchema().dump(category_obj)
+        return CategoryListSchema().dump(category_obj)
 
     @jwt_required()
     def patch(self, category_id):
+        """
+       Updates a new category.
+       ---
+       parameters:
+         - name: category_id
+           in: path
+           required: true
+           type: int
+         - name: name
+           in: request data
+           required: true
+           type: string
+        """
         try:
-            CategoryUpdateSchema().load(request.json)
-            category = Category.query.get_or_404(category_id)
+            CategoryCreateUpdateSchema().load(request.json)
+            category = Category.query.get(category_id)
+            if not category:
+                abort(404, "Invalid Category ID.")
             category.name = request.json.get('name', category.name)
             db.session.commit()
         except ValidationError as err:
             return err.messages, 400
-        return CategorySchema().dump(category)
+        return CategoryListSchema().dump(category)
 
     @jwt_required()
     def delete(self, category_id):
-        category_obj = Category.query.get_or_404(category_id)
+        """
+        Delete a category by ID.
+        ---
+        parameters:
+          - name: category_id
+            in: path
+            description: ID of the category to delete
+            required: true
+            type: integer
+        responses:
+          204:
+            description: No Content.
+          404:
+            description: Invalid category ID.
+        """
+        category_obj = Category.query.get(category_id)
+        if not category_obj:
+            abort(404, "Invalid category ID.")
         db.session.delete(category_obj)
         db.session.commit()
         return {'message': "Category Deleted Successfully"}, 204
@@ -48,17 +93,49 @@ class CategoryResource(Resource):
 class CategoryRetrieveResource(Resource):
     @jwt_required()
     def get(self, category_id):
-        category_obj = Category.query.get_or_404(category_id)
-        return CategorySchema().dump(category_obj)
+        category_obj = Category.query.get(category_id)
+        if not category_obj:
+            abort(404, "Invalid Category ID")
+        return CategoryListSchema().dump(category_obj)
 
 
 class QuestionResource(Resource):
     def get(self):
+        """
+        Get list of all categories
+        """
         questions = Question.query.all()
         return QuestionListSchema(many=True).dump(questions)
 
     @jwt_required()
     def post(self):
+        """
+        Create new question to a category
+        :parameter
+            - name: category_id
+              in: request data
+              type: int
+              required: True
+            - name: question_text
+              in: request data
+              type: String
+              required: True
+            - name: answer
+              in: request data
+              type: String
+              required: True
+            - name: score
+              in: request data
+              type: Int
+              required: True
+
+
+        :responses
+          200:
+            description: Returns the Question that is created
+          404:
+            description: Invalid Question ID
+        """
         try:
             question_obj = QuestionCreateSchema().load(request.json)
         except ValidationError as err:
@@ -70,8 +147,15 @@ class QuestionResource(Resource):
 
     @jwt_required()
     def patch(self, question_id):
+        """
+
+        :param question_id:
+        :return: Updated question object detail
+        """
         new_data = QuestionUpdateSchema().load(request.json, partial=True)
-        question_obj = Question.query.get_or_404(question_id)
+        question_obj = Question.query.get(question_id)
+        if not question_obj:
+            abort(404, "Invalid question ID")
         for key, value in new_data.items():
             setattr(question_obj, key, value)
         db.session.commit()
@@ -79,15 +163,30 @@ class QuestionResource(Resource):
 
     @jwt_required()
     def delete(self, question_id):
+        """
+
+        :param question_id:
+        """
         question_obj = Question.query.get_or_404(question_id)
         db.session.delete(question_obj)
         db.session.commit()
-        return{"message": "Question Deleted successfully."}, 204
+        return {"message": "Question Deleted successfully."}, 204
 
 
 class QuestionDetailResource(Resource):
     @jwt_required()
     def get(self, question_id):
-        question_obj = Question.query.get_or_404(question_id)
-        return QuestionListSchema().dump(question_obj)
+        """
+        Gets a detail of specific Question
+        :param question_id:
+        :return
+            204:
+                description: No Content
+            404:
+                description: Invalid question ID
 
+        """
+        question_obj = Question.query.get(question_id)
+        if not question_obj:
+            abort(404, "Invalid question ID")
+        return QuestionListSchema().dump(question_obj)
